fix: fixKnownBugs
