+++ lib-chat/src/main/java/cn/wzpmc/ChatHandler.java
@@ -10,7 +10,6 @@ import io.netty.handler.logging.LogLevel;
 import io.netty.handler.logging.LoggingHandler;
 import io.netty.handler.stream.ChunkedWriteHandler;
 
-
 /**
  * @Author wzp
  * @Date 2022/1/26
